Extract symbol indexing into perl-symbol-index microcrate
