+++ crates/perl-symbol-index/README.md
@@ -1,0 +1,3 @@
+# perl-symbol-index
+
+Microcrate providing symbol indexing with prefix and fuzzy lookups.
